% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOI.R
\name{add_prefix}
\alias{add_prefix}
\title{Add prefix}
\usage{
add_prefix(df, prefix)
}
\arguments{
\item{df}{data.frame or matrix}

\item{prefix}{a prefix to be added}
}
\value{
matrix / data.frame with new column names.
}
\description{
Adds a prefix to the column names of a matrix
}
