% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_tools.R
\name{geo_dist}
\alias{geo_dist}
\title{Geographical distance}
\usage{
geo_dist(x.loc, z.loc)
}
\arguments{
\item{x.loc}{a 2-length vector. the first value is for longitude,
the second for latitude.}

\item{z.loc}{a 2-length vector. the first value is for longitude,
the second for latitude.}
}
\description{
calculates geo distance between *two* points.
}
