% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOI.R
\name{add_prefix}
\alias{add_prefix}
\title{Add prefix}
\usage{
add_prefix(df, prefix)
}
\arguments{
\item{df}{a data.frame or a matrix.}

\item{prefix}{a prefix to be added.}
}
\value{
a matrix / data.frame with new column names.
}
\description{
Adds a prefix to the column names of a matrix / data.frame.
}
\examples{
X = matrix(rnorm(100), ncol =2)
X = add_prefix(X, "x_")
}
