% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOI.R
\name{OOI}
\alias{OOI}
\title{Outside Option Index}
\usage{
OOI(formula = NULL, X, Z, X.location = NULL, Z.location = NULL,
  wgt = rep(1, nrow(X)), method = "logit", sim.factor = 1,
  dist.fun = geo_dist, dist.order = 2)
}
\arguments{
\item{X}{matrix or data frame with workers characteristics.}

\item{Z}{matrix or data frame with jobs characteristics.}

\item{X.location}{an optional matrix or data frame with location for workers. could be
geographical location (i.e., geo-coordinates) or any other
feature that can be used in order to measure distance between
worker and job using 'dist.fun'.}

\item{Z.location}{same as 'X.location' but for jobs.}

\item{wgt}{an optional vector of weights.}
}
\description{
calculates OOI
}
