% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOI.R
\name{OOI}
\alias{OOI}
\title{Outside Option Index}
\usage{
OOI(formula = NULL, X, Z = NULL, X.location = NULL,
  Z.location = NULL, wgt = rep(1, nrow(X)), pred = T,
  method = "logit", sim.factor = 1, dist.fun = geo_dist,
  dist.order = 2, seed = runif(1, 0, .Machine$integer.max))
}
\arguments{
\item{formula}{a formula describing the model to be fitted in order to
estimate P(Z|X) / P(Z). This formula uses a syntax similar to STATA, and so
"x_" refers to all variables with the prefix "x", while
"z_" refers to all variables with the prefix "z". Similarly, "d" refers
to the distance polynomial (see the example below).}

\item{X}{matrix or data frame with workers characteristics. Note that all column names
should start with "x" (necessary for \code{\link{coef_reshape}}).}

\item{Z}{an optional matrix or data frame with jobs characteristics. Note that all column names
should start with "z" (necessary for \code{\link{coef_reshape}}).}

\item{X.location}{an optional matrix or data frame with location for workers. could be
geographical location (i.e., geo-coordinates) or any other
feature that can be used in order to measure distance between
worker and job using 'dist.fun'.}

\item{Z.location}{same as 'X.location' but for jobs.}

\item{wgt}{an optional vector of weights.}

\item{pred}{logical. If TRUE (default), predicts ooi for the provided data.}

\item{method}{a method for estimating P(Z|X) / P(Z). currently not in use.}

\item{sim.factor}{a variable that determines how much fake data to simulate
(relative to real data).}

\item{dist.fun}{a distance function to calculate the distance between X.location and
Z.location.}

\item{dist.order}{the order of the distance polynomial.}

\item{seed}{the seed of the random number generator.}
}
\value{
an object of class "ooi". This object is a list containing
the following components:
 \item{coeffs}{coefficients from the estimated logit.}
 \item{coeffs_sd}{coefficients standard deviation.}
 \item{pseudo_r2}{McFadden's pseudo-R squared for the estimated logit.}
 \item{standardized_coeffs}{standardized coefficients.}
 \item{ooi}{the outside option index.}
 \item{orig_arg}{a list containing the original arguments (necessary
 for \code{\link{predict.ooi}}).}
}
\description{
calculates the 'outside option index' (defined as
\eqn{-\sum P(Z|X) * log(P(Z|X) / P(Z))})
 for workers, using employer-employee data.
}
