% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOI.R
\name{predict.ooi}
\alias{predict.ooi}
\title{Predict Outside Option Index}
\usage{
\method{predict}{ooi}(
  object,
  new.coef = NULL,
  new.X = NULL,
  new.Z = NULL,
  new.X.location = NULL,
  new.Z.location = NULL,
  new.wgt = NULL
)
}
\arguments{
\item{object}{an ooi object.}

\item{new.coef}{a new *named* vector of coefficients. check the coefficients produced by
the main function to see the right format for this vector.}

\item{new.X}{a new X matrix / data frame.}

\item{new.Z}{a new Z matrix / data frame.}

\item{new.X.location}{a new X.location matrix / data frame.}

\item{new.Z.location}{a new Z.location matrix / data frame.}

\item{new.wgt}{a new vector of weights}
}
\value{
if there are no new arguments, returns the original ooi. otherwise,
returns a vector of ooi calculated using the new arguments.
}
\description{
predicts OOI for new coefficients (for counterfactual analysis)
 and/or new data.
}
