% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOI.R
\name{predict.ooi}
\alias{predict.ooi}
\title{Predict Outside Option Index}
\usage{
\method{predict}{ooi}(
  object,
  new.coef = NULL,
  new.X = NULL,
  new.Z = NULL,
  new.X.location = NULL,
  new.Z.location = NULL,
  new.wgt = NULL,
  hhi = F
)
}
\arguments{
\item{object}{an ooi object.}

\item{new.coef}{a new *named* vector of coefficients. check the coefficients produced by
the main function to see the right format for this vector.}

\item{new.X}{a new X matrix / data frame.}

\item{new.Z}{a new Z matrix / data frame.}

\item{new.X.location}{a new X.location matrix / data frame.}

\item{new.Z.location}{a new Z.location matrix / data frame.}

\item{new.wgt}{a new vector of weights}

\item{hhi}{whether to predict the HHI (Herfindahl-Hirschman Index, an alternative measure for
outside options) instead of the OOI. default is FALSE.}
}
\value{
if there are no new arguments, returns the original results. otherwise,
        returns a vector of ooi/hhi calculated using the new arguments.
}
\description{
predicts the OOI for new coefficients (for counterfactual analysis)
and/or new data.
}
