% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOI.R
\name{predict.ooi}
\alias{predict.ooi}
\title{Predict Outside Option Index}
\usage{
\method{predict}{ooi}(object, new.coef = NULL, new.X = NULL,
  new.Z = NULL, new.X.location = NULL, new.Z.location = NULL,
  new.wgt = NULL)
}
\arguments{
\item{object}{an ooi object.}

\item{new.coef}{new vector of coefficients}
}
\description{
Predict Outside Option Index
}
